Display the project icon in Program service
Part of story #22687 distinguish projects with an icon

The project icon is displayed in the Program breadcrumb.
It is also displayed in the User stories from team when the feature
artifact is expanded.
So make sure to have a feature artifact which is linked with at least a
user story from a team

Change-Id: Ie731518b7bd035577325a1cc1c6c4006b1da65c6