request #23409 Fatal error in program management admin errors presenter
Given I have unchecked the "can submit" permission for the title field of the Releases tracker in a Team project (in a Program Management setup), when I access the Program administration,
I see an error that tells me the current user cannot submit the field. In this error presenter, there is a link. When I click it, I get a fatal error with a LogicException:

LogicException: There are no fields with permissions set for this ugroup
The error appears because the URL is wrong: the URL shows /plugins/tracker/permissions/fields-by-field/917?selected_id=917, but selected_id should be a field identifier, not a tracker identifier.

Change-Id: I7c22d37b9ab01a2a77ff5c6ee12981c243c129f9