Table of Contents can be refreshed
If you try to open a generated document into Microsoft Word or
LibreOffice Writer, if you attemp to refresh the TOC it should
not empty the section.

The issue has been introduced by
20e95e9a3dd343568ba9d9b2a9489f77d645433f where the custom style set on
the artifact titles has been removed. This custom style is necessary,
even if we do not do actual styling with it, because it appears to be
the best way to select the titles we need in the TOC while working as
expected on most of the editors.

Part of story #22569: export `text` fields

Change-Id: Ic4622057be0e6cb1777887e0efcb7da2568c025a