Add a very minimal support of HTML tables
The only goal of this contribution is to have something that is slightly
better than just displaying the text content of the table. Having
something closer to what's rendered in the HTML pages comes with is not
trivial especially when dealing with "complex" DOM elements inside the
table.
Anyway since the tables are not available in the WYSIWYG editor and it
is not something covered by the CommonMark they should not be
encountered too often in artifacts. This basic support level should be
good enough for now.

Part of story #22569: export `text` fields

Change-Id: I31dc95ea1908f43754ebc81c95df968ef066cd2e