Extract hardcoded style options used during the HTML to docx conversion
Some elements were hardcoded during the initial implementation but
should not be the responsability of the transformation process.

The style of the lists as well as which monospace font to use have been
extracted and are now an option of the transformation process. The
unordered list style do not rely on the default provided by the library
anymore to avoid potential changes during an upgrade.

No functional changes.

Part of story #22569: export `text` fields

Change-Id: Id3126b57c975964d986866935f594f97c10716b0