fix: badges does not use the right class in the sidebar
Since tlp-badges in sidebar are on a dark background, we should use CSS
class `tlp-badge-on-dark-background` as per TLP documentation.

It is so-so, since semantic is screwed up (we should not have to update
HTML if a theme decides to use a sidebar in white for example), but
that's what is at our disposition for now. Every usage of this class
should be updated in a dedicated contribution IMHO.

Closes request #23429: Wrong badge color in admin sidebar

Change-Id: I26582838928a089cd00acdf12b08dc4ad022f49d