Build php-mediawiki-tuleap-123 package in the main build pipeline
Can be tested with the tools/rpm/build_and_run_packages.sh script.
You should be able to the the Mediawiki plugin and the
php-mediawiki-tuleap-123 package comes from the local repository (you
might need to add it explictly to the exclusion list of
/etc/yum.repos.d/tuleap-local.repo).

Next time we want to update the php-mediawiki-tuleap-123 package we
will need to update the revision and associated hash to trigger to
get the new package.

Old Debian packaging for the Mediawiki plugin has been removed since it
is not useful.

Part of request #23451: Build the "additional" packages in the main build process

Change-Id: I776309ff6bb9d9d62804691aea0b3389e12a1249