Make sure the additional packages can be rewritten
For now they are copied from the Nix store where they are read-only.
This prevent the build pipeline to work since the package needs to be
rewritten once it has been signed.

To test, you can use the tools/rpm/build_and_run_packages.sh and check
that under /rpms/RPMS/noarch/ the php-mediawiki-tuleap-123 can be
written.

Part of request #23451: Build the "additional" packages in the main build process

Change-Id: I9f97952cade3fde820cbc72c6fd0d298bbf327d6