Build the mediawiki-math-tuleap package in the main build pipeline
To test you can build the package with `nix-build mediawiki-math-tuleap.nix`
and install the resulting RPM inside your dev container. The Mediawiki
Math extension should work.

Note that a change has been made to the package to build statically [0]
to not have references to libraries not available on the targetted OS.

Part of request #23451: Build the "additional" packages in the main build process

[0] https://tuleap.net/plugins/git/tuleap/deps/tuleap/mediawiki-math-tuleap?a=commit&h=729d88339858b94844a52556cfe9d3b4c5a0b7ff

Change-Id: Id73d415e787d3269dac176e14a24224e42ef3324