Add nixpkgs pinned commit date in the Release file of the RPMs
The timestamp is now added to the release field of the packages
containing binaries we built. Since we build them statically it will
allow those packages to be updated when we update our nixpkgs pin.

No functionnal changes expected, the RPMs should still build.

A script has also been added to make it easy to update the pin.

Part of request #23451: Build the "additional" packages in the main build process

Change-Id: I844ccb614b210112238125e2e00fb023a4a0aeac