fix(jira): verification of jira admin is different on Jira Server
The beauty of this patch is that we actually have little to now way to
verify it on Jira Server because the `mypermissions` route behaviour is
not the same between Jira Server and Jira Cloud. So the verification is
really about ensuring that Jira Cloud import still work. And then we
will be able to ask someone to test in the community...

Part of: request #22580 Get list of jira projects doesnt work with Jira Server

Change-Id: Ib9e61fdeeba91a583cd10f8de2e7b1f1f6c84c71