request #23468: Clean-up HTML strings before adding them to the generated documents
To avoid issues with hand crafted/generated by third party HTML content
in the artifacts the strings are now sanitized with DOMPurify. This
should be close to what is done in the different views where HTML
content is displayed.

The one way we know to reproduce this issue is shown as a test case.

Change-Id: I7dc9ba308662a1ef7cafb7157758c9bde8205385