Build the viewvc-theme-tuleap package in the main build pipeline
To test, you can build the package wiht `nix-build
viewvc-theme-tuleap.nix` and install the resulting package in your dev
container. The web browser of a SVN or CVS repository should look the
same.

To fetch the node_modules we use a fixed output derivation (aka FOD) to
execute pnpm. In some Nix circles this is considered as an abused of the
FOD but it is likely to be good enough for us. We do not have to manage
multiple OSes and architectures during the build so it reduces the
chances of something going wrong.

Part of request #23451: Build the "additional" packages in the main build process

Change-Id: Iea79b413bac24c3dec458ad23d13b956f7b6fc53