Fetch the cvs-tuleap and mailman-tuleap packages into the main build pipeline
Rebuilding those packages from scratch is a bit challenging:
* a static build is not so easy since there is direct dependencies to
  parts of the glibc
* the code is not really up to the standards expected by modern build
  tools
* the build process is not really malleable so changing the build tools
  cannot be done without some code changes

Since those packages are not updated often (if at all) the existing RPMs
have been put in the repositories and we use Nix to pull them. This is
still better than what we currently do since we are certain the same
content will be used.

An issue has been identified with the fetchLFS=true feature of nixpkgs'
fetchgit. A local hotfix has been made until a proper fix is upstreamed
and merged.

Part of request #23451: Build the "additional" packages in the main build process

Change-Id: Id1ae64737992dff079416e91b20cf18708a1a56b