chore(jira): Unify detection of JiraCloud vs JiraServer
2 implementations were made, one based on the domain name and another
one based the status of API endpoint. This refactoring aligns on the
second approach as we don't know if domain name is reliable enough.

The refactoring is huge because of the initial design of ClientWrapper
that was embedding the builder in itself and the dependency on the
implementation along the stack.

As we now need to know at build time the type of remote jira server AND
that we have to connect to the target Jira server to detect that, the
ClientWrapperBuilder is now a bit magical because it requires a closure
so the build is done really "just in time".

Part of: request #22580 Get list of jira projects doesnt work with Jira Server

Change-Id: I64cf453771b10d6a042080402c33f04f8c522671