fix: New version should update approval table
Given the platform is configured in siteadmin to "Propose changelog
after new file version drag and drop" in Documents,
And a file item has an approval table in approved state,
When a user drag 'n drop a new version,
Then the modal to set the version name and changelog should ask for
approval table status.

The main change is located in ItemUpdateProperties. The rest of the
change is to make sure that the approval badge is changed after the new
version. For example if the document has approval badge set to
"Approved" and DnD a new version and choose to "Reset the approval
cycle", then after the upload "Not yet" should be displayed in the
approval badge.

Closes request #23481: New version should update approval table

Change-Id: If0fe2828c8adbdf6b6e2e5478754cdfb96bf65a1