Setup infrastructure to add a traceability matrix when exporting step executions
This is the basics of the traceability matrix. For now, even elements
are added without a clear requirements are added to the matrix. Also the
current code is slower than it needs to be because we will call `GET
/testmanagement_executions/:id` twice for each execution.
We consider that anything that point to a test execution is a campaign
because we do not have the information to filter it more precisely. This
approach should however be good enough since the test execution tracker
is mainly a "technical" tracker which is used through the TestManagement
UI. For now we only retrieve the campaign ID.

Since the feature is not really usable outside of a dev env at this time
it is hidden behind a flag. To use it please add the hash `#matrix` to
your URL.

Part of story #23461: have a tracability matrix in my exported document (test executions)

Change-Id: Iac5993f7e8a0fb61235b555f288a866810eb3baa