Deduplicate retrieval of Tracker
part of request #23433 Clean program management tests

No functional change expected.

Given a TrackerIdentifier, we expect to always get a valid Tracker
object, or else to throw an Exception. This behaviour was being repeated
across several adapters. It's easier to reuse it with an interface (in
the Adapters namespace). Adding a Stub lets us reduce the amout of
mocking in our tests.

Change-Id: If919cb1c47469b38faf2b192ab18ff166ce3f98f