Deduplicate retrieval of full Artifact
part of request #23433 Clean program management tests

No functional change expected.

Notes:
There are still usages of \Tracker_ArtifactFactory in adapters:
- calls to getArtifactByIdUserCanView().
This case is different, it also checks user permissions.

- calls to getArtifactById() with an int.
When given an int, it is legitimate to return null. We could call it
with -1 which is never a valid artifact_id. It should not throw an
exception in this case.

Given an ArtifactIdentifier, we expect to always get a valid Artifact
object, or else we throw an Exception. This behaviour was being repeated
across several adapters. It's easier to reuse it with an interface (in
the Adapters namespace). Adding a Stub lets us reduce the amout of
mocking in our tests.

Change-Id: Ia081d58a147f27382da79bb6e0b00d62ba534ba0