Hightlight potentially dangerous bidirectional characters in PR diff
We extend the default CodeMirror regex for the special chars to make
those characters visible. It is not really a great way to make them
visible but it is still better than nothing and if it is good enough to
be a CodeMirror default it is likely good enough for our situation.

Part of request #23496: Add warnings when displaying code with bidirectional Unicode text in Git/PR web UI

Change-Id: I5c93b9e1734f96f543a5bded6e4a19fbf5bfac9e