Fix flaky ProgramPrivacyBuilderTest::testItBuildsProgramPrivacy test
This test is flaky because another test does not clean after itself.
The flaky has also been updated so it does depend on whatever the
current global default is since it is not what is being tested.

The issue is visible with the seed 1636461834 (without having external
plugins in your workspace).

Part of story #21385: display Program's iterations

Change-Id: I33601f9a1191a38d81047cc20553c279f18e54a1