feat: Introduce Parent option in artifact link type
This is a first step to be able to select the Artifact Link type
"Parent" in an artifact.

In order to ease the review, only frontend is updated and this is hidden
behind a feature flag; please run the following command to activate the
feature:
tuleap config-set feature_flag_parent_artlink_type_wip 1

The new option is only available for new artifact link selector.
Existing links will be done later.

If the Artifact Link field already display a Parent selector (part of a
Tracker hierarchy or Program Management enters in the room), then the
new option is not proposed.

Backend to handle the new option will be done in a next contribution.

Starts story #23042: have "parent" as type within artifacts links

Change-Id: I5a1a2196bd7d230db1d9d980a949fb9075cddfb0