feature(jira): Add a configuration switch to force basic auth
As of today Jira Server is always accessed with Personal Access
Token but this cannot work for version older than 8.14

This commit introduce a global switch to force Tuleap to use basic
auth even with Jira Server.

This solution is mainly a workaround until we find a proper way to
address the issue when Tuleap targets one know Jira Server.

Part of: request #24144 Personal Access Token cannot be used with Jira Server < 8.14

Change-Id: I90f81e4c0a4baad33924ffd836f3e9f191b78de1