feat: Begin to export artifact link field
This is part of story #22663 export `artifact links` field

This is the first step of artifact links field export in the generated
document. For now, only the artifact ID and the type are exported.

The artifact title is not yet provided in the 2 sub titles.

To enable the feature, #artlinks must be added to the report URL.

Change-Id: Icdab6bd2623345a5509e867e384a406b256285bc