chore: Cover getFieldDataFromRESTValue with unit tests
In order to be more confident in next refactoring, we should ensure that
this method is well covered.

Extracting the method to a dedicated object was considered but would
involve too much unwanted change. This forces us to introduce dummy
method to set a dao to ease test, due to crappy initial design. Cleaning
up this mess is not the main objective, at least we have more tests for
next contributions.

No functional changes are expected.

Part of story #23042: have "parent" as type within artifacts links

Change-Id: I7613c77c535e80c5f4a098e691045453a4171653