request #24156: Make sure commit signature based on S/MIME and SSH can be extracted
This is a technical change to make clear in the code that we might
encounter signatures that are not GPG based.
No functional changes expected.

If you want to try to sign commits with an SSH key you need to have
access to Git 2.34.0. If it is not yet available on your system, you can
use the following Nix derivation to get it:

    { pkgs ? import <nixpkgs> {} }:
    pkgs.git.overrideAttrs (oldAttrs: rec {
      version = "2.34.0";
      src = pkgs.fetchurl {
        url = "https://www.kernel.org/pub/software/scm/git/git-${version}.tar.xz";
        sha256 = "07s1c9lzlm4kpbb5lmxy0869phg7037pv4faz5hlqyb5csrbjv7x";
      };
    })

Change-Id: If4d57e3c1d85423fd6b376a2981e393114e16f33