fix: improve "more than one parent" message
When an artifact has more than one parent, currently there are two
different warnings displayed:
* A regular feedback one saying "The artifact A has more than one parent: B, C"
* A bootstrap alert sàying "The artifact has more than one parent. Cannot display rest of hierarchy."

Two concerns are raised about this situation.
1. The bootstrap alert lacks margin, especially visible when project has
a background
2. There is a repetition "The artifact has more than one parent"

This issue is here since long time ago, but might be highlighted with
the new ability to select Parent type for a bunch of artifacts ids to
link.

At first, we might want to merge the two sentences: "The artifact A has
more than one parent: B, C. Cannot display rest of hierarchy." However
the first feedback is raised too deep in the call stack, in a factory
(sic) and is probably displayed in pages where we don't display the
hierarchy but need the information about possibly weird situation.

Extracting the message out of the factory will probably require too much
effort compared to the scope of the current story and what is proposed
here:
1. Adjust the second warning to hide the repetition
2. Treat the second warning as a regular feedback

This give something like:
> The artifact A has more than one parent: B, C
> When more than one parent, we cannot display rest of hierarchy.

You can test it by linking A as "Child" in B and C.
Or you can link B and C as "Parent" in A after enabling the feature flag:
tuleap config-set feature_flag_parent_artlink_type_wip 1

Part of story #23042: have "parent" as type within artifacts links

Change-Id: I8a7f4107794541946ead753fc5b048cfd095388a