chore: rename ArtifactLink\Nature namespace to ArtifactLink\Type
Since the renaming of "nature" to "type" [five years ago][1], only the front
end and various places in backend have been changed to reflect that.

This means that today we have code that deal with both "nature" and
"type" terms. This is confusing, we should only have "type".

This is a first contribution to clean this up.

Part of story #23042: have "parent" as type within artifacts links

[1]: See git #tuleap/stable/db50cfab0a5084ce75e7d6055bf1395a8590a41a

Change-Id: Ibf28d1398af3de8c87f18c3b55bd62ef9af21306