refacto(program-management) Redirections to program_management app
This refactoring is needed before introducing the redirections to the
planned-iterations app after an iteration creation/update.

The main issues with the existing code were:
- Duplication of checks in the pre and post artifact creation/edition hooks
- The code in the hooks callbacks are specific to the program_management app.
  It does not ease the handling of the redirections specific to the
  iteration creations/updates.

It appeared difficult to have hexagonal architecture here since we are
in a context of tracker events. If you have a better idea to make the
target code better, I'll be glad to consider it.

-- No functional change expected

part of: story #21385 display Program's iterations

Change-Id: Ieedcb578d33a678d777f60c798c988c784807599