chore: mass rename of Nature to Type in tracker
I'm sorry in advance for the lengthy review; I started by renaming stuff
in Type creation/edition and ended up renaming mostly everything in
tracker and few plugins.

No functional changes is expected. Non exhaustive list of impacted
features:
* CRUD artifact link types in site admin » tracker
* Tracker main administration
* Report (add/remove some dedicated artifact link types column)
* Artifact view (artifact link field, follow-up comments details)
* Children tab of artifact view

REST representations have been left as is to not introduce a breaking
change. The same for XML export/import. Maybe an approach would be to
introduce a new "type" entry alongside "nature" so that we can deprecate
the latter. This can be done later.

The report's session is also left intact to not force users to be
disconnected. It can also be done later if needed.

Part of story #23042: have "parent" as type within artifacts links

Change-Id: I87567a5e3b325f25592ec30cb3601454762ad4b8