feat(jira): Introduce JiraServer ChangeLog
In order to be able to reproduce the behaviour of a JiraServer we are
able to capture the logs with the approriate feature flag. This patch
introduce the structure of a a high level test of Jira import that
loads the payloads returned by an actual Jira Server (data where
anonymized). In it's current form, it is disabled because that would
require to support the whole Jira Import and the goal of this change is
to make smaller increment.

This increment focuses on the management of ChangeLog.

Part of: request #22580 Creation of new tracker from Jira doesn't work with Jira Server

Change-Id: I3fbc65a1e392cee5f20e4eae94d94d357a81bdb9