feat(jira): Introduce Jira ReplayImport command
In addition to tests, it's useful to be able to import data that were
generated by the Tuleap debug capture on another server. The main use
case is debug and verification because it's not simple to find the right
assertion to make on the generated XML without having the actual result
and compare.

In order to be able to test the command (and make the CI happy) the
import of comments was disabled for JiraServer (it crashes anyway).

If you want to run some tests manually you can do it with:

    tuleap import-project:replay-jira \
        --path plugins/tracker/tests/unit/Tracker/Creation/JiraImporter/Import/_fixtures/SBX \
        --project=one_tuleap_project \
        --user=one_tuleap_user \
        --tracker-name=bug0015

Part of: request #22580 Creation of new tracker from Jira doesn't work with Jira Server

Change-Id: Ibb7339717a313e0559f2c5b743628d51036515a3