chore: move image-loader in dedicated lib
In order to avoid DRY between TestPlan and DocGen, we should have common
stuff in a dedicated library. This will be mostly useful for artifact
generation. As a baby step, this contribution only export the image
loader.

No functional changes are expected:
* Test plan document has a cover page with the platform logo
* Tracker report document has a cover page with the platform logo and
images in the exported artifacts

Part story #23462: export test plan as a docx document

Change-Id: I3ba1f0d4e2fb821412fbf3212bb67d10b4baab9a