Build the API Explorer with Vite
The main goal of this contribution is to make possible to bundle app
with stylesheets with Vite.

Vite expects the main entrypoint to define the stylesheets to load so it
excludes them from the manifets files as root entries. This is fine for
most cases where we can tell the app to load the stylesheets.
There is a few cases in Tuleap where this is not possible (mostly global
themes/"framework") which we will need additional work to be built with
Vite. There is however no need at this point but it could be handled in
a similar way than what vite_ruby does with a custom manifest plugin [0].

The warning that is now visible when buiding the API Explorer is due to
a IE7 workaround and will be removed the next time we update swagger-ui [1]
(and they rebuild their distributed files).

No functionnal changes expected.

Part of request #23415: Allow to bundle app with Vite

[0] https://github.com/ElMassimo/vite_ruby/blob/vite-plugin-ruby%403.0.3/vite-plugin-ruby/src/manifest.ts
[1] https://github.com/swagger-api/swagger-ui/pull/7526

Change-Id: I5e6b13756d3c4fbff99439b470fdb795a8ab7b72