feat: docx is hidden behind a feature flag
Since the feature is far from complete for the upcoming release, we need
to hide it under a feature flag.

Go to TestPlan interface. => You can only download a spreadsheet.
Add `#feature-flag-docx` in the url and refresh the page => you can
download both spreadsheet and document.

A missing test has been added for ExportButton.

No test is added for the feature flag since it is only an `indexOf` that
will be removed later.

Part story #23462: export test plan as a docx document

Change-Id: Ib2ca9b1da6c9ca3deb8bf4697b88ebc05176f1f1