feat: Display artifacts content in test report
Instead of displaying a list of artifact titles for the backlog section,
we now display the full artifact, like it is done in tracker report
document generation.

If backlog items have artifact link to another backlog itemin the
generated document, then an anchor is used to facilitate navigation.

No functional change in tracker report document generation.

Part of story #23462: export test plan as a docx document

Change-Id: I650d0e1b55d8dee571728bd89d29ee54353a4d00