rename ContentStore -> SearchFeatures
part of story #21385 display Program's iterations

No functional change expected when planning a Feature in a Program
Increment or in the Program Increment's backlog REST route.

Notes:
The interface returned tracker_name which was never used, so it was
removed.
The interface and adapter were in the "Feature/Content" namespace which
was incorrect. They have been moved to "ProgramIncrement/Content"
because they return Features (which are the PI's content), not User
Stories (which are the Feature's content).
The interface has been hardened so it requires a
ProgramIncrementIdentifier instead of any int.

Change-Id: I665a2f2bd169a736bcb4410bb8c5b223bf92e314