Merge RetrieveTrackerFromUserStory in TrackerOfArtifactRetriever
part of story #21385 display Program's iterations

No functional change expected in the REST routes that return User
Stories:
- GET Iteration content
- GET Feature content (in Program app, when you browse the children of a
  feature)

Notes:
Retrieving the Tracker of an Artifact is always the same process. It is
the same for a Timebox or a User Story or a Feature, so we aim to have
one adapter that implements many interfaces, similar to
RetrieveCrossReference or RetrieveTitle.

Change-Id: Ieea479453d827f715e47ffece036ef97b7b44ee6