Init Project Sidebar custom element
This contribution inits the project sidebar custom elements.
It can be built with `pnpm run build` and a dev server can be started
with `pnpm run dev`.

Since the goal is to provide a package that is directly usable, we
bundle everything in the distributed bundles, that's why Vue is marked
as a dev dependencies there is no need for users to pull it when they
`npm install` the package.

Part of request #23465 Publish a project sidebar custom element

Change-Id: I845097f56eb209a7324312f8c3aaca304a216ad9