Display the project name in the project sidebar custom element
This contribution allows to setup major parts of the element
infrastructure: it pulls the CSS and adds unit tests.

Note that the CSS does not goes through PostCSS Autoprefixer. This will be
managed for all Vite configurations in a independant contributions.

You can test the change by going into the folder and running `npm run
dev`.

Part of request #23465 Publish a project sidebar custom element

Change-Id: I6b3ecf66faf71386db24c4aef4877a0a39053217