Drop PlannableFeatureIdentifier
part of story #21385 display Program's iterations

Expected functional change: when user cannot see a feature and tries to
GET its children user stories, the message will change
from: "You can not see this feature."
to: "Could not find feature with id #123"

No other functional change expected when getting the User Stories of a
Feature

Notes:
In order to retrieve the Program Increment backlog, I need to get all
the User Stories of all the Features planned in that PI. To do that, I
needed to pass a PlannableFeatureIdentifier.

PlannableFeatureIdentifier was performing an additional check compared
with FeatureIdentifier. But when you have a collection of
FeatureIdentifier, that check becomes redundant (see below).

Instead, we should move that check in FeatureIdentifier. Only `fromId()`
is changed.

- `buildCollectionFromProgramIncrement()` does not need to double-check
because the SQL query from ContentDao ensures that it returns plannable
features
- same for `buildCollectionFromProgram()`: the SQL query from
FeaturesDao ensures that it returns plannable features
- `fromIdAndProgram()` should also do that check. It is used in more
"sensitive" parts of the plugin, so I think we should do another
contribution for it.

Since the REST handler relied on specific exceptions for certain error
codes (400 vs 404), the FeatureChecker throws exceptions instead of
returning a boolean.

Change-Id: I6945dd6d88c109b20188101e00a2b720c8364f88