chore: extract retrieval of executions
In order to be able to build the "tests" section of the TestPlan report
document, we need to retrieve all executions of the campaigns of the
milestone.

However this retrieval is already done to build the traceability matrix.
To save resources and the rainforest, we should only retrieve executions
once.

This commit only deal with the extract. Use of executions to build the
tests section will be done later. No functional changes expected.

Bonus: not needed for the current refacto but `tests` property has been
added to `ExportDocument` interface while we were here.

Part of story #23462: export test plan as a docx document

Change-Id: I1a944d6d9430816b77db44a1680d9299456ecfb2