Add dedicated interface for User Stories visibility
part of story #21385 display Program's iterations

No functional change expected in the following use cases:
- GET the content of an Iteration
- GET the content of a Program increment
- GET the Program backlog

Notes:
"ArtifactIsVisible" is too generic. Artifact is a Tracker plugin
concept, we should have specific interfaces for "Program Management"
concepts such as Feature or User Story. The adapter does the same thing
so it's ok if it implements a bunch of similar interfaces.

I took the opportunity to simplify SearchChildrenOfFeature so it returns
int[] instead of a DAO row array. It also takes a FeatureIdentifier
instead of any int.

I also took the opportunity to simplify FeatureHasUserStoriesVerifier,
which was moved in the Domain since it only depends on Domain
interfaces.

Change-Id: Iaf7b3349803fac8dadd33a2dbab8f2056d77f67f