feat: artifact content in tests section
Retrieve artifact content of each test involved in the campaigns of the
release and display it in the "tests" section.

If a backlog item is linked to a test, and both are part of the
generated document, then there are anchors in the artifact link sections
of both items so that navigation is facilitated.

The test/step results of each tests will be done later.

Part of story #23462: export test plan as a docx document

Change-Id: Ic0519a48f24b142922d3fe8be056df2a90058815