chore(jira): Refactor how fields are imported
Sorry about the giantic review, I didn't manage to slim it more as we
need to transform the way Tuleap XML is built for fields and it's quite
central to the import process.

In this refactoring I tried to limit as much as possible the changes so
basic things were not changed like the name of
`ContainersXMLCollectionBuilder` class that no longer build a collection
of containers but I feel that adding a proper rename on top of that
would be a nightmare for review.

The keys to understand the refactoring:
- First we need it because we will have to distinguish in future the
  fields that can be seen at submit time and fields that can be seen at
  update time only. With previous code (before this commit) the
  transformation done in `JiraToTuleapFieldTypeMapper` to directly get
  bare XML made impossible to carry information about the permissions
  (handled way after in the PermissionXMLExporter without connection
  with actual JiraField).
  By using the XMLTracker & XMLField objects we introduce an
  intermidiate representation that will allow to introduce what we want
  (in that case, permissions). However at this stage, to limit the
  refactoring, permissions are still managed by the original code.
- `ContainersXMLCollectionBuilder` was merged with `AlwaysThereFieldsExporter`
- FieldMapperXml is gone (replaced by XMLTracker->export()) but we still
  need to generate the XML after creation of fields otherwise all the
  reports and semantic stuff should be converted to XMLTracker way of
  doing the things and that would be too much for one commit.
  In the meantime I dropped usage of FieldName formatter because it
  seems that field names are always compliant with Tuleap norms because
  they are either hard coded (so we can check easily) or they are custom
  and jira custom fields are always custom_XXXX where XXXX is a number.
  As a matter of fact it seems name formatter was only needed for cross
  reference field whose pseudo name (to match jira stuff) was made with
  dot. I choose to renamed the constant value at source instead.

The tricky part is about the list values (like status or select box
options) because in the previous implementation the XML id (for Tuleap)
was actually generated when the value was fetched from Jira aka
JiraFieldAPIAllowedValueRepresentation. As field values, at Tuleap side,
are now created when the XMLListField is created (see
JiraToTuleapFieldTypeMapper for select fields for instance) we have to
re-create the values so they match the actual value ids in FieldMappingCollection.

This is far from perfect but, again, it's a tradeoff to limit the impact
of this refactoring.

Part of: request #24222 Jira import fields that are only updated

Change-Id: If0f919904d0262cfab2013ab4cc1459280045ca1