feat(jira): import fields used at edition
This commit allow jira import to take into account fields that can only
be set at issue update. Until now, those fields were ignored because we
relied on createmeta route only.

With this implementation, all (supported) fields should be imported but
there is one caveat: the will all be "Submittable" in Tuleap. Depending
on the trackers that might generate a very large artifact creation form.
The management of the appropriate permissions will be done in a
dedicated commit.

Part of: request #24222 Jira import fields that are only updated

Change-Id: Ib564cf533b48b04020c694d2a7afc8d66fa3ed0b