chore(jira): Integration tests now have a baseline for reference
The JiraXMLExporterTest allows to verify that the import sequence is
working well but until now it was not that much useful in pure test as
tests themselves where rather simple (only asserting the number of
artifacts). In the following stories we are going to work on deeper
changes in the way import is done so we could leverage a better
integration test.

This commit introduces two things
- A test based on Jira Cloud (in completment of the Jira Server fixtures
  we already have). As it's exported from Enalean test instance, I
  choose to leave it almost un modified as there are no confidential
  material here.
- A baseline of generated tracker for each fixture so we can detect when
  we are changing what we export.

Part of: request #24222 Jira import fields that are only updated

Change-Id: I8ea35fdfbd515b843096aa3b3d81ad64f042d4b5